+++ tests/alarm_handler/test_pointing_data_alarm.py
@@ -9,6 +9,7 @@ from tests.resources.test_harness.helpers import retry_tango_command
 from tests.resources.test_support.constant import alarm_handler1
 
 
+@pytest.mark.batch2
 @pytest.mark.SKA_mid
 @scenario(
     "../features/test_harness/xtp_50437_nan_alarm.feature",