+++ tests/dish_vcc_configuration/test_xtp-44894_kval_error_scenario.py
@@ -15,6 +15,7 @@ from tests.resources.test_harness.utils.common_utils import JsonFactory
 from tests.resources.test_support.common_utils.result_code import ResultCode
 
 
+@pytest.mark.batch1
 @pytest.mark.SKA_mid
 @scenario(
     "../features/dish_vcc_initialization/xtp_44894_kvalue_validation.feature",