+++ tests/resources/test_harness/central_node_mid.py
@@ -1,7 +1,6 @@
 import json
 import logging
 import os
-import time
 from typing import List, Tuple
 
 from ska_control_model import ObsState, ResultCode
@@ -454,8 +453,6 @@ class CentralNodeWrapperMid(CentralNodeWrapper):
         Args:
             input_string (str): Release resource input json
         """
-        time.sleep(3)
-
         result, message = self.central_node.ReleaseResources(input_string)
         return result, message