+++ tests/tmc_harness_testcases/bdd_scenarios/test_xtp_49344_recover_subarray_stuck_in_resourcing_with_defective_sdp.py
@@ -1,9 +1,15 @@
 """Test module to test Assignresources while SDP is defective."""
+
 import pytest
 from pytest_bdd import given, parsers, scenario
 from ska_control_model import ObsState
+from ska_tango_testing.mock.placeholders import Anything
 
 from tests.resources.test_harness.central_node_mid import CentralNodeWrapperMid
+from tests.resources.test_harness.constant import (
+    RESET_DEFECT,
+    SDP_BACK_TO_INITIAL_STATE,
+)
 from tests.resources.test_harness.event_recorder import EventRecorder
 from tests.resources.test_harness.helpers import (
     get_device_simulators,
@@ -13,6 +19,7 @@ from tests.resources.test_harness.simulator_factory import SimulatorFactory
 from tests.resources.test_harness.utils.common_utils import JsonFactory
 
 
+@pytest.mark.batch1
 @pytest.mark.SKA_mid
 @scenario(
     "../features/test_harness/"
@@ -34,6 +41,7 @@ def telescope_is_in_resourcing_obsstate(
     central_node_mid: CentralNodeWrapperMid,
     event_recorder: EventRecorder,
     command_input_factory: JsonFactory,
+    simulator_factory: SimulatorFactory,
     subarray_id: str,
 ):
     """A method to check if telescope in is resourcing obsSstate."""
@@ -44,16 +52,22 @@ def telescope_is_in_resourcing_obsstate(
     )
     # Induce fault on SDP Subarry so that it raises exception and
     # returns to the obsState EMPTY
+    _, sdp_sim, _, _, _, _ = get_device_simulators(simulator_factory)
+    sdp_sim.SetDefective(SDP_BACK_TO_INITIAL_STATE)
     assign_input_json = prepare_json_args_for_centralnode_commands(
-        "assign_resources_mid_invalid_sdp_resources", command_input_factory
+        "assign_resources_mid", command_input_factory
     )
 
-    central_node_mid.perform_action("AssignResources", assign_input_json)
+    pytest.command_result = central_node_mid.perform_action(
+        "AssignResources", assign_input_json
+    )
 
 
 @given("SDP subarray raised error goes back to obsState EMPTY")
 def sdp_subarray_stuck_is_in_empty(
-    event_recorder: EventRecorder, simulator_factory: SimulatorFactory
+    central_node_mid: CentralNodeWrapperMid,
+    event_recorder: EventRecorder,
+    simulator_factory: SimulatorFactory,
 ):
     "Method to check SDP subarray is in EMPTY."
     _, sdp_sim, _, _, _, _ = get_device_simulators(simulator_factory)
@@ -63,6 +77,13 @@ def sdp_subarray_stuck_is_in_empty(
         "obsState",
         ObsState.EMPTY,
     )
+    event_recorder.has_change_event_occurred(
+        central_node_mid.central_node,
+        "longRunningCommandResult",
+        (pytest.command_result[1][0], Anything),
+        lookahead=15,
+    )
+    sdp_sim.SetDefective(RESET_DEFECT)
 
 
 # from conftest.py