+++ tests/tmc_harness_testcases/test_subarray_negative_scenarios.py
@@ -19,6 +19,7 @@ from tests.resources.test_harness.utils.enums import SimulatorDeviceType
 
 
 class TestSubarrayNodeNegative(object):
+    @pytest.mark.batch2
     @pytest.mark.SKA_mid
     def test_subarray_assign_csp_unresponsive(
         self,
@@ -67,6 +68,7 @@ class TestSubarrayNodeNegative(object):
 
         csp_sim.SetDefective(RESET_DEFECT)
 
+    @pytest.mark.skip(reason="Duplicate scenario")
     @pytest.mark.SKA_mid
     def test_subarray_configure_when_csp_stuck_in_configuring(
         self,
@@ -110,6 +112,7 @@ class TestSubarrayNodeNegative(object):
             csp_sim, "Configure", csp_input_json
         )
 
+    @pytest.mark.skip(reason="Duplicate scenario")
     @pytest.mark.SKA_mid
     def test_subarray_configure_when_sdp_stuck_in_configuring(
         self,
@@ -151,6 +154,7 @@ class TestSubarrayNodeNegative(object):
         )
         assert len(get_recorded_commands(sdp_sim)) == 1
 
+    @pytest.mark.batch2
     @pytest.mark.SKA_mid
     def test_subarray_configure_when_dish_stuck_in_slew(
         self,
@@ -171,8 +175,8 @@ class TestSubarrayNodeNegative(object):
         subarray_node.move_to_on()
         subarray_node.force_change_of_obs_state("IDLE")
 
-        # Dish master should go to slew in no more than 0.1 sec
-        pointing_state_duration_params = '[["SLEW",0.1]]'
+        # Dish master should go to Ready in no more than 0.1 sec
+        pointing_state_duration_params = '[["READY",0.1]]'
         dish_sim.AddTransition(pointing_state_duration_params)
 
         subarray_node.execute_transition("Configure", argin=input_json)
@@ -185,52 +189,3 @@ class TestSubarrayNodeNegative(object):
                 subarray_node.subarray_node, "obsState", ObsState.READY
             )
         assert device_received_this_command(dish_sim, "ConfigureBand1", "True")
-
-    # As per Initial analysis, push_obs_state_event() method required to set
-    # obsState is not present in HelperSubArrayDevice, hence obsState of
-    # CspSubarray is not getting updated and test fails.
-    @pytest.mark.skip(reason="Fails in assertions after Fault")
-    @pytest.mark.SKA_mid
-    def test_subarray_configure_when_csp_goes_to_fault_then_ready(
-        self,
-        subarray_node,
-        command_input_factory,
-        simulator_factory,
-        event_recorder,
-    ):
-        input_json = prepare_json_args_for_commands(
-            "configure_mid", command_input_factory
-        )
-        csp_input_json = prepare_json_args_for_commands(
-            "csp_configure_mid", command_input_factory
-        )
-        csp_sim = simulator_factory.get_or_create_simulator_device(
-            SimulatorDeviceType.MID_CSP_DEVICE
-        )
-
-        event_recorder.subscribe_event(subarray_node.subarray_node, "obsState")
-        event_recorder.subscribe_event(csp_sim, "obsState")
-        subarray_node.move_to_on()
-        subarray_node.force_change_of_obs_state("IDLE")
-
-        # CSP should go to configuring in no more than 0.1 sec
-        obs_state_duration_params = '[["FAULT",0.1],["READY",0.1]]'
-        csp_sim.AddTransition(obs_state_duration_params)
-
-        subarray_node.execute_transition("Configure", argin=input_json)
-
-        assert event_recorder.has_change_event_occurred(
-            csp_sim, "obsState", ObsState.FAULT
-        )
-        assert event_recorder.has_change_event_occurred(
-            csp_sim, "obsState", ObsState.READY
-        )
-        assert event_recorder.has_change_event_occurred(
-            subarray_node.subarray_node, "obsState", ObsState.CONFIGURING
-        )
-        assert event_recorder.has_change_event_occurred(
-            subarray_node.subarray_node, "obsState", ObsState.READY
-        )
-        assert device_received_this_command(
-            csp_sim, "Configure", csp_input_json
-        )