+++ tests/tmc_harness_testcases/test_subarray_obs_transitions.py
@@ -24,6 +24,7 @@ class TestSubarrayNodeObsStateTransitions(object):
             # ("READY", "End", "IDLE"),
         ],
     )
+    @pytest.mark.batch2
     @pytest.mark.SKA_mid
     def test_subarray_obs_transitions_valid_data(
         self,
@@ -103,6 +104,7 @@ class TestSubarrayNodeObsStateTransitions(object):
             obs_state=destination_obs_state, timeout=expected_timeout_sec
         )
 
+    @pytest.mark.batch2
     @pytest.mark.SKA_mid
     @pytest.mark.parametrize(
         "source_obs_state, trigger, args_for_command,\