+++ tests/tmc_robustness/Configure_failure_handling_tests/test_configure_handling_on_csp_obstate_idle_failure.py
@@ -6,16 +6,21 @@ from ska_control_model import ObsState
 from tango import DevState
 
 from tests.resources.test_harness.constant import (
-    COMMAND_FAILED_WITH_EXCEPTION_OBSSTATE_IDLE,
+    ABORT_COMPLETED,
+    COMMAND_COMPLETED,
+    COMMAND_FAILED_WITH_EXCEPTION_OBSSTATE_CONFIGURING_IDLE,
 )
 from tests.resources.test_harness.helpers import (
+    check_for_device_command_event,
     get_device_simulators,
     prepare_json_args_for_centralnode_commands,
     prepare_json_args_for_commands,
 )
 from tests.resources.test_harness.utils.enums import SimulatorDeviceType
+from tests.resources.test_support.enum import PointingState
 
 
+@pytest.mark.batch1
 @pytest.mark.SKA_mid
 @scenario(
     "../features/xtp-28436.feature",
@@ -38,6 +43,12 @@ def test_configure_handling_on_csp_subarray_obsstate_idle_failure():
 def given_tmc(central_node_mid, subarray_node, event_recorder):
     event_recorder.subscribe_event(
         central_node_mid.central_node, "telescopeState"
+    )
+    event_recorder.subscribe_event(
+        central_node_mid.central_node, "longRunningCommandResult"
+    )
+    event_recorder.subscribe_event(
+        subarray_node.subarray_node, "longRunningCommandResult"
     )
     event_recorder.subscribe_event(subarray_node.subarray_node, "obsState")
     central_node_mid.move_to_on()
@@ -67,11 +78,26 @@ def given_tmc_subarray_assign_resources(
     assign_input_json = prepare_json_args_for_centralnode_commands(
         "assign_resources_mid", command_input_factory
     )
-    central_node_mid.perform_action("AssignResources", assign_input_json)
+    _, unique_id = central_node_mid.perform_action(
+        "AssignResources", assign_input_json
+    )
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node,
         "obsState",
         ObsState.IDLE,
+    )
+    assert event_recorder.has_change_event_occurred(
+        central_node_mid.central_node,
+        "longRunningCommandResult",
+        (unique_id[0], COMMAND_COMPLETED),
+    )
+
+    assert check_for_device_command_event(
+        subarray_node.subarray_node,
+        "longRunningCommandResult",
+        COMMAND_COMPLETED,
+        event_recorder,
+        "AssignResources",
     )
 
 
@@ -88,7 +114,10 @@ def given_tmc_subarray_configure_is_in_progress(
 ):
     csp_sim, _, _, _, _, _ = get_device_simulators(simulator_factory)
     csp_sim.SetDefective(
-        json.dumps(COMMAND_FAILED_WITH_EXCEPTION_OBSSTATE_IDLE)
+        json.dumps(COMMAND_FAILED_WITH_EXCEPTION_OBSSTATE_CONFIGURING_IDLE)
+    )
+    sdp_sim = simulator_factory.get_or_create_simulator_device(
+        SimulatorDeviceType.MID_SDP_DEVICE
     )
     configure_input_json = prepare_json_args_for_commands(
         "configure_mid", command_input_factory
@@ -96,6 +125,16 @@ def given_tmc_subarray_configure_is_in_progress(
     subarray_node.execute_transition("Configure", configure_input_json)
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node,
+        "obsState",
+        ObsState.CONFIGURING,
+    )
+    assert event_recorder.has_change_event_occurred(
+        sdp_sim,
+        "obsState",
+        ObsState.CONFIGURING,
+    )
+    assert event_recorder.has_change_event_occurred(
+        csp_sim,
         "obsState",
         ObsState.CONFIGURING,
     )
@@ -133,11 +172,25 @@ def csp_subarray_returns_to_obsstate_idle(event_recorder, simulator_factory):
 @given(
     parsers.parse("the TMC SubarrayNode {subarray_id} stucks in CONFIGURING")
 )
-def given_tmc_subarray_stuck_configuring(subarray_node, simulator_factory):
+def given_tmc_subarray_stuck_configuring(
+    central_node_mid, subarray_node, simulator_factory, event_recorder
+):
     csp_sim = simulator_factory.get_or_create_simulator_device(
         SimulatorDeviceType.MID_CSP_DEVICE
     )
     assert subarray_node.subarray_node.obsState == ObsState.CONFIGURING
+
+    for dish_id in ["SKA001", "SKA036", "SKA063", "SKA100"]:
+        event_recorder.subscribe_event(
+            central_node_mid.dish_leaf_node_dict[dish_id], "pointingState"
+        )
+        assert event_recorder.has_change_event_occurred(
+            central_node_mid.dish_leaf_node_dict[dish_id],
+            "pointingState",
+            PointingState.TRACK,
+            lookahead=15,
+        )
+
     # Disable CSP Subarray fault
     csp_sim.SetDefective(json.dumps({"enabled": False}))
 
@@ -148,7 +201,7 @@ def given_tmc_subarray_stuck_configuring(subarray_node, simulator_factory):
     )
 )
 def send_command_abort(subarray_node, event_recorder):
-    subarray_node.execute_transition("Abort", argin=None)
+    _, pytest.unique_id = subarray_node.execute_transition("Abort", argin=None)
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node,
         "obsState",
@@ -161,7 +214,9 @@ def send_command_abort(subarray_node, event_recorder):
         "the SDP subarray {subarray_id} transitions to obsState ABORTED"
     )
 )
-def sdp_subarray_transitions_to_aborted(simulator_factory, event_recorder):
+def sdp_subarray_transitions_to_aborted(
+    simulator_factory, event_recorder, subarray_node
+):
     sdp_sim = simulator_factory.get_or_create_simulator_device(
         SimulatorDeviceType.MID_SDP_DEVICE
     )
@@ -200,6 +255,11 @@ def tmc_subarray_transitions_to_aborted(subarray_node, event_recorder):
         ObsState.ABORTED,
         lookahead=9,
     )
+    assert event_recorder.has_change_event_occurred(
+        subarray_node.subarray_node,
+        "longRunningCommandResult",
+        (pytest.unique_id[0], ABORT_COMPLETED),
+    )
 
 
 @when(
@@ -208,7 +268,9 @@ def tmc_subarray_transitions_to_aborted(subarray_node, event_recorder):
     )
 )
 def send_command_restart(subarray_node, event_recorder, simulator_factory):
-    subarray_node.execute_transition("Restart", argin=None)
+    _, pytest.restart_unqiue_id = subarray_node.execute_transition(
+        "Restart", argin=None
+    )
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node,
         "obsState",
@@ -259,6 +321,11 @@ def tmc_subarray_transitions_to_empty(subarray_node, event_recorder):
         "obsState",
         ObsState.EMPTY,
     )
+    assert event_recorder.has_change_event_occurred(
+        subarray_node.subarray_node,
+        "longRunningCommandResult",
+        (pytest.restart_unqiue_id[0], COMMAND_COMPLETED),
+    )
 
 
 @then(
@@ -273,19 +340,42 @@ def configure_executed_on_subarray(
     assign_input_json = prepare_json_args_for_centralnode_commands(
         "assign_resources_mid", command_input_factory
     )
-    central_node_mid.perform_action("AssignResources", assign_input_json)
+    _, unique_id = central_node_mid.perform_action(
+        "AssignResources", assign_input_json
+    )
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node,
         "obsState",
         ObsState.IDLE,
     )
+    assert event_recorder.has_change_event_occurred(
+        central_node_mid.central_node,
+        "longRunningCommandResult",
+        (unique_id[0], COMMAND_COMPLETED),
+    )
+
+    assert check_for_device_command_event(
+        subarray_node.subarray_node,
+        "longRunningCommandResult",
+        COMMAND_COMPLETED,
+        event_recorder,
+        "AssignResources",
+    )
     configure_input_json = prepare_json_args_for_commands(
         "configure_mid", command_input_factory
     )
-    subarray_node.execute_transition("Configure", configure_input_json)
+    _, unique_id = subarray_node.execute_transition(
+        "Configure", configure_input_json
+    )
 
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_node,
         "obsState",
         ObsState.READY,
+    )
+
+    assert event_recorder.has_change_event_occurred(
+        subarray_node.subarray_node,
+        "longRunningCommandResult",
+        (unique_id[0], COMMAND_COMPLETED),
     )